add methods to compute score

+++ scrabble-score/src/test/java/ScrabbleScoreTest.java
@@ -5,43 +5,43 @@ public class ScrabbleScoreTest {
   @Test
   public void calculateScore_forWordWithOnlyOnePointValueLetters_three() {
     ScrabbleScore testScrabbleScore = new ScrabbleScore();
-    assertEquals(3, testScrabbleScore.calculateScore("run"));
+    assertEquals((Integer)3, testScrabbleScore.calculateScore("run"));
   }
 
   @Test
   public void calculateScore_forWordWithOneOrTwoPointValueLetters_five() {
     ScrabbleScore testScrabbleScore = new ScrabbleScore();
-    assertEquals(5, testScrabbleScore.calculateScore("dog"));
+    assertEquals((Integer)5, testScrabbleScore.calculateScore("dog"));
   }
 
   @Test
   public void calculateScore_forWordWithOneTwoOrThreePointValueLetters_seven() {
     ScrabbleScore testScrabbleScore = new ScrabbleScore();
-    assertEquals(7, testScrabbleScore.calculateScore("pigs"));
+    assertEquals((Integer)7, testScrabbleScore.calculateScore("pigs"));
   }
 
   @Test
   public void calculateScore_forWordWithOneTwoThreeOrFourPointValueLetters_twelve() {
     ScrabbleScore testScrabbleScore = new ScrabbleScore();
-    assertEquals(12, testScrabbleScore.calculateScore("piggy"));
+    assertEquals((Integer)12, testScrabbleScore.calculateScore("piggy"));
   }
 
   @Test
   public void calculateScore_forWordWithOneTwoThreeFourOrFivePointValueLetters_fourteen() {
     ScrabbleScore testScrabbleScore = new ScrabbleScore();
-    assertEquals(14, testScrabbleScore.calculateScore("viking"));
+    assertEquals((Integer)14, testScrabbleScore.calculateScore("viking"));
   }
 
   @Test
   public void calculateScore_forWordWithOneTwoThreeFourFiveOrEightPointValueLetters_nineteen() {
     ScrabbleScore testScrabbleScore = new ScrabbleScore();
-    assertEquals(19, testScrabbleScore.calculateScore("jumpy"));
+    assertEquals((Integer)19, testScrabbleScore.calculateScore("jumpy"));
   }
 
   @Test
   public void calculateScore_forWordWithAnyPointValueLetters_twentyFive() {
     ScrabbleScore testScrabbleScore = new ScrabbleScore();
-    assertEquals(25, testScrabbleScore.calculateScore("quickly"));
+    assertEquals((Integer)25, testScrabbleScore.calculateScore("quickly"));
   }
 }
